alzkb/requirements.txt
google-genai
pandas
tqdm
python-dotenv
rdflib
scipy
numpy
streamlit
graphviz
pandas
networkx
requests
